refactor: move @common components to @ui
