+++ frontend/src/components/modals/GamesHistoryModal.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
-import BaseModal from "@components/common/BaseModal";
 import { GameSummary } from "@components/stats/History";
+import BaseModal from "@ui/BaseModal";
 import PaginationActions from "@ui/PaginationActions";
 
 const GamesHistoryModal = ({
