+++ frontend/src/pages/settings.tsx
@@ -1,5 +1,7 @@
 import MainLayout from "@components/layout";
+
 import Image from "next/image";
+
 import Avatar from "/public/images/default-avatar.jpg";
 
 export default function SettingsPage() {
@@ -102,9 +104,9 @@ export default function SettingsPage() {
                         <input
                           type="checkbox"
                           value=""
-                          className="sr-only peer"
+                          className="peer sr-only"
                         />
-                        <div className="w-11 h-6 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-blue-300 dark:peer-focus:ring-blue-800 rounded-full peer dark:bg-gray-700 peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all dark:border-gray-600 peer-checked:bg-blue-600"></div>
+                        <div className="peer w-11 h-6 bg-gray-200 peer-focus:outline-none peer-focus:ring-4 peer-focus:ring-blue-300 dark:peer-focus:ring-blue-800 rounded-full dark:bg-gray-700 peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-[2px] after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all dark:border-gray-600 peer-checked:bg-blue-600"></div>
                         <span className="ml-3 text-sm font-medium text-gray-900 dark:text-gray-700">
                           Activate 2FA
                         </span>
